1. Seperate Classes into single file. 2. Add commandlines to control nano gpt running mode. 3. Add nano-gpt-cpu & nano-gpt-gpu. 4. Add nano-gpt-cli-cpu & nano-gpt-cli-gpu. 4. Seperate nano-gpt-base.
