+++ nano-gpt-base/Enums.cs
