+++ nano-gpt-base/HelperFunctions.cs
