+++ nano-gpt-base/TokenEncoder.cs
