+++ nano-gpt-base/input.txt
