+++ nano-gpt-cli-cpu/nano-gpt-cli-cpu.csproj
@@ -3,14 +3,13 @@
     <PropertyGroup>
         <OutputType>Exe</OutputType>
         <TargetFramework>net7.0</TargetFramework>
-        <RootNamespace>nano_gpt</RootNamespace>
+        <RootNamespace>nano_gpt_cli_cpu</RootNamespace>
         <ImplicitUsings>enable</ImplicitUsings>
         <Nullable>disable</Nullable>
         <OutputPath>build/</OutputPath>
     </PropertyGroup>
 
     <ItemGroup>
-        <PackageReference Include="System.CommandLine" Version="2.0.0-beta1.21308.1" />
-        <ProjectReference Include="..\nano-gpt\nano-gpt.csproj" />
+        <ProjectReference Include="..\nano-gpt-cpu\nano-gpt-cpu.csproj" />
     </ItemGroup>
 </Project>
