+++ nano-gpt-cli-gpu/nano-gpt-cli-gpu.csproj
@@ -1,0 +1,15 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+    <PropertyGroup>
+        <OutputType>Exe</OutputType>
+        <TargetFramework>net7.0</TargetFramework>
+        <RootNamespace>nano_gpt_cli_gpu</RootNamespace>
+        <ImplicitUsings>enable</ImplicitUsings>
+        <Nullable>disable</Nullable>
+        <OutputPath>build/</OutputPath>
+    </PropertyGroup>
+
+    <ItemGroup>
+        <ProjectReference Include="..\nano-gpt-gpu\nano-gpt-gpu.csproj" />
+    </ItemGroup>
+</Project>
