+++ nano-gpt-cpu/NanoGptGenerating.cs
@@ -1,6 +1,6 @@
 using Shared;
 using TorchSharp;
-using NanoGptSetting;
+using NanoGptArgsSettings;
 using Settings = NanoGptSetting.Settings;
 using Gpt;
 using FileLib;
@@ -12,9 +12,9 @@ namespace NanoGptGenerating
         public static void Generate(TokenEncoder tokenEncoder, int vocabSize, ArgsSettings argsSettings)
         {
             GptLanguageModel model = new GptLanguageModel("My_Language_Model", vocabSize).to(Settings.Device);
-            if (FileUtils.FileExists(argsSettings.SaveLocation(vocabSize)))
+            if (FileUtils.FileExists(argsSettings.SaveLocation(vocabSize, Settings.SettingsKey)))
             {
-                model.load(argsSettings.SaveLocation(vocabSize));
+                model.load(argsSettings.SaveLocation(vocabSize, Settings.SettingsKey));
             }
             model.GenerateAndPrint(tokenEncoder, int.MaxValue);
         }
