+++ nano-gpt-cpu/README.md
